requirements.txt
unittest2>=1.1.0
SQLAlchemy>=1.0.11
PyMySQL>=0.7.9
python-keystoneclient>=2.3.1
python-glanceclient>=2.3.0
python-neutronclient>=5.1.0
python-novaclient==2.29.0
python-cinderclient==1.6.0
#keystoneauth1==2.12.1
#Pyro4==4.49
dill==0.2.4
Flask==0.11.1
Flask-RESTful==0.3.5
mock==2.0.0
nose==1.3.7
setuptools==28.3.0